New instantiation API:
Changes:
- The parameters for every Trait now will be delivered in an array.
- Optional Parameter-Arrays now need to be delivered and can not be omitted in the instantiation parameter list if they are in before or in the middle of other required parameters.
Supply either `undefined` or a `[]` here.
- Parameters can be ommitted if all Sub-Traits have all-optional or non-existing parameters.

Reason for this change:
- We can not reliably determine the parameter count of a class passed to `FusionOf`at runtime.
We previously relied on `<constructor>.length` to give us the parameter count and used the special case `0` to determine that we can omit supplying parameters to it.
This does not hold true in case a default constructor is in place (a derived subclass with no `constructor` member defined). In this case the `length` is always 0 independent of the base class' constructor parameter count.